update ui admin page

+++ src/main/java/com/service/Controller/ComplaintServlet.java
@@ -7,6 +7,7 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 
 import java.io.IOException;
 import java.util.List;
@@ -14,12 +15,14 @@ import java.util.List;
 @WebServlet("/submitComplaint")
 
 public class ComplaintServlet extends HttpServlet {
+
     CompliantDao complainDao = new CompliantDao();
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-        List<Complain> complains = complainDao.getAllComplaints();
+        String id = request.getParameter("id");
+        List<Complain> complains = complainDao.getComplaintsByUserId(id);
         request.setAttribute("complaints", complains);
         request.getRequestDispatcher("/Complain.jsp").forward(request, response);
 
@@ -83,18 +86,23 @@ public class ComplaintServlet extends HttpServlet {
          String id = req.getParameter("id");
          String userId = req.getParameter("userId");
          String subject = req.getParameter("subject");
+         String status = req.getParameter("status");
          String description = req.getParameter("description");
 
-        Complain complain = new Complain(id ,userId,subject, description, "pending", "none","Not Work");
+        if (!status.equals("Pending")) {
+            response.getWriter().println("<script>alert('You cannot update complaint in " + status + " status!'); window.location.href='submitComplaint';</script>");
+            return;
+        }
+
+        Complain complain = new Complain(id, userId, subject, description, status, "none", "Not Work");
 
         boolean result = complainDao.updateCompliant(complain);
         response.setContentType("text/html");
 
         if (result) {
-            System.out.println("updated");
+            response.getWriter().println("<script>alert('Complaint updated successfully!'); window.location.href='submitComplaint';</script>");
         } else {
-            response.getWriter().println("<h3>complain update failed!</h3>");
-
+            response.getWriter().println("<script>alert('Complaint update failed!'); window.location.href='submitComplaint';</script>");
         }
 
     }
@@ -102,13 +110,19 @@ public class ComplaintServlet extends HttpServlet {
 
     public void deleteComplain(HttpServletRequest req, HttpServletResponse response) throws IOException {
         String id = req.getParameter("id");
+        String status = req.getParameter("status");
+
+        response.setContentType("text/html");
+
+        if (!status.equals("Pending")) {
+            response.getWriter().println("<script>alert('You cannot delete complaint in " + status + " status!'); window.location.href='submitComplaint';</script>");
+            return;
+        }
         boolean result = complainDao.deleteComplain(id);
         if (result) {
-            System.out.println("updated");
+            response.getWriter().println("<script>alert('Complaint deleted successfully!'); window.location.href='submitComplaint';</script>");
         } else {
-            response.getWriter().println("<h3>complain update failed!</h3>");
-
+            response.getWriter().println("<script>alert('Complaint deletion failed!'); window.location.href='submitComplaint';</script>");
         }
-        response.setContentType("text/html");
     }
 }
